requirements.txt
flask
google-generativeai
fpdf
bs4
serpapi